pyproject.toml
[tool.poetry]
name = "axaxax"
version = "0.1.0"
description = ""
authors = ["Hexvel <dimaabdukarimov1234@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
tortoise-orm = "^0.21.6"
aerich = "^0.7.2"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
aiomysql = "^0.2.0"
vkbottle = "^4.4.2"
